+++ src/Catalog/Condition.php
@@ -13,8 +13,16 @@ use MyCLabs\Enum\Enum;
  * @method static Condition BRAND_NEW()
  * @method static Condition USED()
  */
-final class Condition extends Enum
+final class Condition extends Enum implements \JsonSerializable
 {
     private const BRAND_NEW = 'N';
     private const USED = 'U';
+
+    /**
+     * @inheritdoc
+     */
+    public function jsonSerialize(): string
+    {
+        return $this->getValue();
+    }
 }